Changed unsigned to uint32_t in many places

+++ src/wwd.cpp
@@ -19,29 +19,29 @@ unsigned wap_wwd_get_checksum(const wap_wwd *wwd)
     return wwd->checksum;
 }
 
-size_t wap_wwd_get_plane_count(const wap_wwd *wwd)
+unsigned wap_wwd_get_plane_count(const wap_wwd *wwd)
 {
 	return wwd->planes.size();
 }
 
-int wap_wwd_set_plane_count(wap_wwd *wwd, size_t count)
+int wap_wwd_set_plane_count(wap_wwd *wwd, unsigned count)
 {
     return wap::handle_exceptions([=](){
         wwd->planes.resize(count);
     });
 }
 
-wap_plane *wap_wwd_get_plane(wap_wwd *wwd, size_t plane_index)
+wap_plane *wap_wwd_get_plane(wap_wwd *wwd, unsigned plane_index)
 {
 	return &wwd->planes[plane_index];
 }
 
-size_t wap_wwd_get_tile_description_count(const wap_wwd *wwd)
+unsigned wap_wwd_get_tile_description_count(const wap_wwd *wwd)
 {
     return wwd->tile_descriptions.size();
 }
 
-wap_tile_description *wap_wwd_get_tile_description(wap_wwd *wwd, size_t description_index)
+wap_tile_description *wap_wwd_get_tile_description(wap_wwd *wwd, unsigned description_index)
 {
     return &wwd->tile_descriptions[description_index];
 }
@@ -67,7 +67,7 @@ int wap_plane_set_map_dimensions(wap_plane *plane, unsigned w, unsigned h)
     
 }
 
-unsigned wap_plane_get_tile(const wap_plane *plane, size_t x, size_t y)
+unsigned wap_plane_get_tile(const wap_plane *plane, unsigned x, unsigned y)
 {
     return plane->tiles[y * plane->tiles_wide + x];
 }
@@ -77,43 +77,43 @@ void wap_plane_set_tile(wap_plane *plane, unsigned x, unsigned y, unsigned tile)
     plane->tiles[y * plane->tiles_wide + x] = tile;
 }
 
-size_t wap_plane_get_image_set_count(const wap_plane *plane)
+unsigned wap_plane_get_image_set_count(const wap_plane *plane)
 {
 	return plane->image_sets.size();
 }
 
-int wap_plane_set_image_set_count(wap_plane *plane, size_t count)
+int wap_plane_set_image_set_count(wap_plane *plane, unsigned count)
 {
     return wap::handle_exceptions([=](){
         plane->image_sets.resize(count);
     });
 }
 
-const char *wap_plane_get_image_set(const wap_plane *plane, size_t image_set_index)
+const char *wap_plane_get_image_set(const wap_plane *plane, unsigned image_set_index)
 {
 	return plane->image_sets[image_set_index].c_str();
 }
 
-int wap_plane_set_image_set(wap_plane *plane, size_t image_set_index, const char *image_set)
+int wap_plane_set_image_set(wap_plane *plane, unsigned image_set_index, const char *image_set)
 {
     return wap::handle_exceptions([=](){
         plane->image_sets[image_set_index].assign(image_set);
     });
 }
 
-size_t wap_plane_get_object_count(const wap_plane *plane)
+unsigned wap_plane_get_object_count(const wap_plane *plane)
 {
     return plane->objects.size();
 }
 
-int wap_plane_set_object_count(wap_plane *plane, size_t count)
+int wap_plane_set_object_count(wap_plane *plane, unsigned count)
 {
     return wap::handle_exceptions([=](){
         plane->objects.resize(count);
     });
 }
 
-wap_object *wap_plane_get_object(wap_plane *plane, size_t object_index)
+wap_object *wap_plane_get_object(wap_plane *plane, unsigned object_index)
 {
     return &plane->objects[object_index];
 }
